Changed the project structure and added spring boot application
Changed the project structure and added spring boot application
