+++ maker-checker-dashboard/backend/src/main/java/com/makerchecker/MakerCheckerDashboard.java
@@ -1,4 +1,4 @@
-package com.example.makerchecker;
+package com.makerchecker;
 
 
 import org.springframework.boot.SpringApplication;
