--- maker-checker-dashboard/backend/src/main/java/com/makerchecker/ProductSubmission.java
@@ -1,14 +1,0 @@
-package com.example.makerchecker;
-
-public class ProductSubmission {
-    private String product_code;
-    private Double rate_of_intrest;
-    private Long maker_id;
-
-    public String getProduct_code() { return product_code; }
-    public void setProduct_code(String product_code) { this.product_code = product_code; }
-    public Double getRate_of_intrest() { return rate_of_intrest; }
-    public void setRate_of_intrest(Double rate_of_intrest) { this.rate_of_intrest = rate_of_intrest; }
-    public Long getMaker_id() { return maker_id; }
-    public void setMaker_id(Long maker_id) { this.maker_id = maker_id; }
-}
