+++ maker-checker-dashboard/backend/src/main/java/com/makerchecker/controller/ProductController.java
@@ -1,6 +1,8 @@
 
-package com.example.makerchecker;
+package com.makerchecker.controller;
 
+import com.makerchecker.service.ProductService;
+import com.makerchecker.model.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
@@ -13,17 +15,17 @@ public class ProductController {
     private ProductService productService;
 
     @PostMapping("/submit")
-    public void submitProduct(@RequestBody ProductSubmission submission) {
+    public void submitProduct(@RequestBody Product.ProductSubmission submission) {
         productService.submitProduct(submission);
     }
 
     @GetMapping("/maker/{makerId}")
-    public List<ProductApprovalView> getMakerProducts(@PathVariable Long makerId) {
+    public List<Product.ProductApprovalView> getMakerProducts(@PathVariable Long makerId) {
         return productService.getMakerProducts(makerId);
     }
 
     @GetMapping("/checker/{checkerId}")
-    public List<ProductApprovalView> getCheckerPending(@PathVariable Long checkerId) {
+    public List<Product.ProductApprovalView> getCheckerPending(@PathVariable Long checkerId) {
         return productService.getCheckerPending(checkerId);
     }
