+++ maker-checker-dashboard/backend/src/main/java/com/makerchecker/model/Approval.java
@@ -1,4 +1,4 @@
-package com.example.makerchecker;
+package com.makerchecker.model;
 import javax.persistence.*;
 
 @Entity
