+++ maker-checker-dashboard/backend/src/main/java/com/makerchecker/model/CheckerDecision.java
@@ -1,4 +1,4 @@
-package com.example.makerchecker;
+package com.makerchecker.model;
 
 public class CheckerDecision {
     private Long approval_id;
