+++ maker-checker-dashboard/backend/src/main/java/com/makerchecker/repository/ApprovalRepository.java
@@ -1,4 +1,4 @@
-package com.example.makerchecker;
+package com.makerchecker.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
